Improved PyTorch cnn | Checked permutation invariance | Other fixes
