+++ genobuilder.py
@@ -182,10 +182,16 @@ class Genobuilder:
         rng = random.Random(seed)
 
         for p in param_vals:
-          sims.append(msprime.simulate(
-              sample_size=self.num_samples, Ne=self.params['Ne'].val, length=self.seq_len,
-              mutation_rate=self.params['mu'].val, recombination_rate=p, random_seed=seed))
-
+            sims.append(
+                msprime.simulate(
+                    sample_size=self.num_samples,
+                    Ne=self.params["Ne"].val,
+                    length=self.seq_len,
+                    mutation_rate=self.params["mu"].val,
+                    recombination_rate=p,
+                    random_seed=seed,
+                )
+            )
 
         mat = np.zeros((self.num_reps, self.num_samples, self.fixed_dim))
 
@@ -197,7 +203,6 @@ class Genobuilder:
         mat = np.expand_dims(mat, axis=1)
 
         return mat
-
 
     def simulate_msprime(self, params, randomize=False, proposals=False):
         """Simulate demographic data, returning a tensor with n_reps number
@@ -788,7 +793,7 @@ if __name__ == "__main__":
     args = parser.parse_args()
     params_dict = OrderedDict()
 
-    params_dict["r"] = Parameter("r", 1.25e-9, 1e-10, (1e-12, 1e-7), inferable=True)
+    params_dict["r"] = Parameter("r", 1.25e-9, 1e-8, (1e-11, 1e-7), inferable=True)
     params_dict["mu"] = Parameter("mu", 1.25e-8, 1e-9, (1e-11, 1e-7), inferable=False)
     params_dict["Ne"] = Parameter("Ne", 10000, 14000, (5000, 15000), inferable=True)
